renamed auth folder in tests to mock
