+++ backend/codingchallenge/api/tests/mock/mockAuth.py
