+++ backend/codingchallenge/api/tests/views/application/test_editApplication.py
@@ -5,7 +5,7 @@ from django.core import serializers
 from rest_framework import status
 from rest_framework.test import APITestCase
 
-from ...auth.mockAuth import MockAuth
+from ...mock.mockAuth import MockAuth
 from ....models.application import Application
 from ....models.challenge import Challenge
