+++ backend/codingchallenge/api/tests/views/challenge/test_putChallengeAdmin.py
@@ -1,7 +1,7 @@
 from rest_framework import status
 from rest_framework.test import APITestCase
 
-from ...auth.mockAuth import MockAuth
+from ...mock.mockAuth import MockAuth
 from ....models.challenge import Challenge
 from ....serializers import GetChallengeSerializer
